+++ modules/lambdas/main.tf
@@ -297,6 +297,72 @@ resource "aws_lambda_function" "request_status_for_granule" {
   }
 }
 
+# API Gateway
+resource "aws_api_gateway_rest_api" "request_status_for_granule_api" {
+  name = "${var.prefix}_request_status_for_granule_api"
+}
+
+resource "aws_api_gateway_resource" "request_status_for_granule_api_resource" {
+  path_part   = "{proxy+}"
+  parent_id   = aws_api_gateway_rest_api.request_status_for_granule_api.root_resource_id
+  rest_api_id = aws_api_gateway_rest_api.request_status_for_granule_api.id
+}
+
+resource "aws_api_gateway_method" "request_status_for_granule_api_method" {
+  rest_api_id   = aws_api_gateway_rest_api.request_status_for_granule_api.id
+  resource_id   = aws_api_gateway_resource.request_status_for_granule_api_resource.id
+  http_method   = "ANY"
+  # todo: Make sure this is locked down against external access.
+  authorization = "NONE"
+}
+
+resource "aws_api_gateway_integration" "request_status_for_granule_api_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.request_status_for_granule_api.id
+  resource_id             = aws_api_gateway_resource.request_status_for_granule_api_resource.id
+  http_method             = aws_api_gateway_method.request_status_for_granule_api_method.http_method
+  integration_http_method = "POST"
+  type                    = "AWS"
+  uri                     = aws_lambda_function.request_status_for_granule.invoke_arn
+}
+
+resource "aws_api_gateway_method_response" "request_status_for_granule_response_200" {
+  rest_api_id = aws_api_gateway_rest_api.request_status_for_granule_api.id
+  resource_id = aws_api_gateway_resource.request_status_for_granule_api_resource.id
+  http_method = aws_api_gateway_method.request_status_for_granule_api_method.http_method
+  status_code = "200"
+}
+
+resource "aws_api_gateway_integration_response" "request_status_for_granule_api_response" {
+  depends_on = [aws_api_gateway_integration.request_status_for_granule_api_integration]
+  rest_api_id = aws_api_gateway_rest_api.request_status_for_granule_api.id
+  resource_id = aws_api_gateway_resource.request_status_for_granule_api_resource.id
+  http_method = aws_api_gateway_method.request_status_for_granule_api_method.http_method
+  status_code = aws_api_gateway_method_response.request_status_for_granule_response_200.status_code
+
+  # Transforms the backend JSON response to XML
+  response_templates = {
+    "application/json" = <<EOF
+#set($inputRoot = $input.path('$'))
+$input.json("$")
+#if($input.path("stackTrace") != '')
+#set($context.responseOverride.status = 500)
+#elseif($input.path("httpStatus") != '')
+#set($context.responseOverride.status = $input.path('httpStatus'))
+#end
+EOF
+  }
+}
+
+resource "aws_lambda_permission" "request_status_for_granule_api_permission" {
+  statement_id  = "AllowExecutionFromAPIGateway"
+  action        = "lambda:InvokeFunction"
+  function_name = aws_lambda_function.request_status_for_granule.function_name
+  principal     = "apigateway.amazonaws.com"
+
+  # More: http://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-control-access-using-iam-policies-to-invoke-api.html
+  # source_arn = "arn:aws:execute-api:${var.region}:${var.accountId}:${aws_api_gateway_rest_api.request_status_for_granule_api.id}/*/${aws_api_gateway_method.request_status_for_granule_api_method.http_method}${aws_api_gateway_resource.request_status_for_granule_api_resource.path}"
+}
+
 
 # request_status_for_job - Provides recovery status information for a job.
 # ==============================================================================
@@ -326,6 +392,73 @@ resource "aws_lambda_function" "request_status_for_job" {
     }
   }
 }
+
+# API Gateway
+resource "aws_api_gateway_rest_api" "request_status_for_job_api" {
+  name = "${var.prefix}_request_status_for_job_api"
+}
+
+resource "aws_api_gateway_resource" "request_status_for_job_api_resource" {
+  path_part   = "{proxy+}"
+  parent_id   = aws_api_gateway_rest_api.request_status_for_job_api.root_resource_id
+  rest_api_id = aws_api_gateway_rest_api.request_status_for_job_api.id
+}
+
+resource "aws_api_gateway_method" "request_status_for_job_api_method" {
+  rest_api_id   = aws_api_gateway_rest_api.request_status_for_job_api.id
+  resource_id   = aws_api_gateway_resource.request_status_for_job_api_resource.id
+  http_method   = "ANY"
+  # todo: Make sure this is locked down against external access.
+  authorization = "NONE"
+}
+
+resource "aws_api_gateway_integration" "request_status_for_job_api_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.request_status_for_job_api.id
+  resource_id             = aws_api_gateway_resource.request_status_for_job_api_resource.id
+  http_method             = aws_api_gateway_method.request_status_for_job_api_method.http_method
+  integration_http_method = "POST"
+  type                    = "AWS"
+  uri                     = aws_lambda_function.request_status_for_job.invoke_arn
+}
+
+resource "aws_api_gateway_method_response" "request_status_for_job_response_200" {
+  rest_api_id = aws_api_gateway_rest_api.request_status_for_job_api.id
+  resource_id = aws_api_gateway_resource.request_status_for_job_api_resource.id
+  http_method = aws_api_gateway_method.request_status_for_job_api_method.http_method
+  status_code = "200"
+}
+
+resource "aws_api_gateway_integration_response" "request_status_for_job_api_response" {
+  depends_on = [aws_api_gateway_integration.request_status_for_job_api_integration]
+  rest_api_id = aws_api_gateway_rest_api.request_status_for_job_api.id
+  resource_id = aws_api_gateway_resource.request_status_for_job_api_resource.id
+  http_method = aws_api_gateway_method.request_status_for_job_api_method.http_method
+  status_code = aws_api_gateway_method_response.request_status_for_job_response_200.status_code
+
+  # Transforms the backend JSON response to XML
+  response_templates = {
+    "application/json" = <<EOF
+#set($inputRoot = $input.path('$'))
+$input.json("$")
+#if($input.path("stackTrace") != '')
+#set($context.responseOverride.status = 500)
+#elseif($input.path("httpStatus") != '')
+#set($context.responseOverride.status = $input.path('httpStatus'))
+#end
+EOF
+  }
+}
+
+resource "aws_lambda_permission" "request_status_for_job_api_permission" {
+  statement_id  = "AllowExecutionFromAPIGateway"
+  action        = "lambda:InvokeFunction"
+  function_name = aws_lambda_function.request_status_for_job.function_name
+  principal     = "apigateway.amazonaws.com"
+
+  # More: http://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-control-access-using-iam-policies-to-invoke-api.html
+  # source_arn = "arn:aws:execute-api:${var.region}:${var.accountId}:${aws_api_gateway_rest_api.request_status_for_job_api.id}/*/${aws_api_gateway_method.request_status_for_job_api_method.http_method}${aws_api_gateway_resource.request_status_for_job_api_resource.path}"
+}
+
 
 # post_copy_request_to_queue - Posts to two queues for notifying copy_files_to_archive lambda and updating the DB."
 # ==============================================================================
@@ -407,6 +540,72 @@ resource "aws_lambda_function" "orca_catalog_reporting_dummy" {
     subnet_ids         = var.lambda_subnet_ids
     security_group_ids = [module.lambda_security_group.vpc_postgres_ingress_all_egress_id]
   }
+}
+
+# API Gateway
+resource "aws_api_gateway_rest_api" "orca_catalog_reporting_api" {
+  name = "${var.prefix}_orca_catalog_reporting_api"
+}
+
+resource "aws_api_gateway_resource" "orca_catalog_reporting_api_resource" {
+  path_part   = "{proxy+}"
+  parent_id   = aws_api_gateway_rest_api.orca_catalog_reporting_api.root_resource_id
+  rest_api_id = aws_api_gateway_rest_api.orca_catalog_reporting_api.id
+}
+
+resource "aws_api_gateway_method" "orca_catalog_reporting_api_method" {
+  rest_api_id   = aws_api_gateway_rest_api.orca_catalog_reporting_api.id
+  resource_id   = aws_api_gateway_resource.orca_catalog_reporting_api_resource.id
+  http_method   = "ANY"
+  # todo: Make sure this is locked down against external access.
+  authorization = "NONE"
+}
+
+resource "aws_api_gateway_integration" "orca_catalog_reporting_api_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.orca_catalog_reporting_api.id
+  resource_id             = aws_api_gateway_resource.orca_catalog_reporting_api_resource.id
+  http_method             = aws_api_gateway_method.orca_catalog_reporting_api_method.http_method
+  integration_http_method = "POST"
+  type                    = "AWS"
+  uri                     = aws_lambda_function.orca_catalog_reporting_dummy.invoke_arn
+}
+
+resource "aws_api_gateway_method_response" "orca_catalog_reporting_response_200" {
+  rest_api_id = aws_api_gateway_rest_api.orca_catalog_reporting_api.id
+  resource_id = aws_api_gateway_resource.orca_catalog_reporting_api_resource.id
+  http_method = aws_api_gateway_method.orca_catalog_reporting_api_method.http_method
+  status_code = "200"
+}
+
+resource "aws_api_gateway_integration_response" "orca_catalog_reporting_api_response" {
+  depends_on = [aws_api_gateway_integration.orca_catalog_reporting_api_integration]
+  rest_api_id = aws_api_gateway_rest_api.orca_catalog_reporting_api.id
+  resource_id = aws_api_gateway_resource.orca_catalog_reporting_api_resource.id
+  http_method = aws_api_gateway_method.orca_catalog_reporting_api_method.http_method
+  status_code = aws_api_gateway_method_response.orca_catalog_reporting_response_200.status_code
+
+  # Transforms the backend JSON response to XML
+  response_templates = {
+    "application/json" = <<EOF
+#set($inputRoot = $input.path('$'))
+$input.json("$")
+#if($input.path("stackTrace") != '')
+#set($context.responseOverride.status = 500)
+#elseif($input.path("httpStatus") != '')
+#set($context.responseOverride.status = $input.path('httpStatus'))
+#end
+EOF
+  }
+}
+
+resource "aws_lambda_permission" "orca_catalog_reporting_api_permission" {
+  statement_id  = "AllowExecutionFromAPIGateway"
+  action        = "lambda:InvokeFunction"
+  function_name = aws_lambda_function.orca_catalog_reporting_dummy.function_name
+  principal     = "apigateway.amazonaws.com"
+
+  # More: http://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-control-access-using-iam-policies-to-invoke-api.html
+  # source_arn = "arn:aws:execute-api:${var.region}:${var.accountId}:${aws_api_gateway_rest_api.request_status_for_granule_api.id}/*/${aws_api_gateway_method.request_status_for_granule_api_method.http_method}${aws_api_gateway_resource.request_status_for_granule_api_resource.path}"
 }